Integrate Phys_Math into Physics
Physics.cc was the only user of Phys_Math, combining
these files gives the compiler better optimizing
opportunities and removes two fairly empty source
files.
