+++ src/game/TileEngine/Physics.cc
@@ -79,6 +79,56 @@ constexpr float EPSILONP = 0.01f;
 
 #define REALOBJ2ID(o) 				((o) - ObjectSlots)
 
+namespace {
+vector_3 VAdd( vector_3 *a, vector_3 *b )
+{
+	vector_3 c;
+
+	c.x = a->x + b->x;
+	c.y = a->y + b->y;
+	c.z = a->z + b->z;
+
+	return( c );
+}
+
+vector_3 VMultScalar( vector_3 *a, float b )
+{
+	vector_3 c;
+
+	c.x = a->x * b;
+	c.y = a->y * b;
+	c.z = a->z * b;
+
+	return( c );
+}
+
+
+float VDotProduct( vector_3 *a, vector_3 *b )
+{
+	return ( ( a->x * b->x ) + ( a->y * b->y ) + ( a->z * b->z ) );
+}
+
+
+vector_3 VGetNormal( vector_3 *a )
+{
+	vector_3 c;
+	const float length = VDotProduct(a, a);
+	if (length == 0)
+	{
+		c.x = 0;
+		c.y = 0;
+		c.z = 0;
+	}
+	else
+	{
+		const float OneOverLength = 1 / sqrt(length);
+		c.x = OneOverLength * a->x;
+		c.y = OneOverLength * a->y;
+		c.z = OneOverLength * a->z;
+	}
+	return ( c );
+}
+}
 
 /// OBJECT POOL FUNCTIONS
 static REAL_OBJECT* GetFreeObjectSlot(void)
@@ -995,9 +1045,7 @@ static BOOLEAN PhysicsCheckForCollisions(REAL_OBJECT* pObject, INT32* piCollisio
 
 		if ( fDoCollision )
 		{
-			pObject->CollisionNormal.x		= vTemp.x;
-			pObject->CollisionNormal.y		= vTemp.y;
-			pObject->CollisionNormal.z		= vTemp.z;
+			pObject->CollisionNormal = vTemp;
 			pObject->CollisionElasticity  = dElasity;
 			pObject->iOldCollisionCode  = iCollisionCode;
 
@@ -1716,7 +1764,6 @@ void CalculateLaunchItemParamsForThrow(SOLDIERTYPE* const pSoldier, INT16 sGridN
 	vector_3 vForce, vDirNormal;
 	INT16    sFinalGridNo;
 	BOOLEAN  fArmed = FALSE;
-	UINT16   usLauncher;
 	INT16    sStartZ;
 	INT8     bMinMissRadius, bMaxMissRadius, bMaxRadius;
 	FLOAT    fScale;
